refactor(notifications): reorganize supabase function files
