--- functions-unused/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "module": "esnext",
-    "moduleResolution": "NodeNext",
-    "lib": ["esnext"],
-    "types": ["@cloudflare/workers-types"],
-    "esModuleInterop": true
-  }
-}
